Improve how border and padding are clipped; show error messages for unsupported values (#143)
You can check the tests to know more about this change. 

Previously we are only using `clip-path` for overflow but it's not that
descriptive (e.g. when there're complex borders), so here we add `mask`
to handle overflow all together. Also fixed how content is clipped when
there are border radius, border and/or padding.